Add files via upload
This is a file with some features created using the existing vars, and modeling done for log/reg as a base model and oversampling method applied to log/reg. I used a dataset, for which I dropped all the zeroes. Next time, try with all zeroes and with imputed values.